More create-release-pr changes
